PercentDisCountPolicy which extends DiscountPolicy

+++ src/main/java/_chapter02/Money.java
@@ -32,7 +32,7 @@ public class Money {
     return new Money(this.amount.subtract(amount.amount));
   }
 
-  public Money times(int percent) {
+  public Money times(double percent) {
     return new Money(this.amount.multiply(BigDecimal.valueOf(percent)));
   }
